MAINT: rename feed to drives
