+++ hnn_core/__init__.py
@@ -1,5 +1,5 @@
 from .dipole import simulate_dipole, read_dipole, average_dipoles
-from .feed import feed_event_times
+from .drives import feed_event_times
 from .params import Params, read_params
 from .network import Network, CellResponse, read_spikes
 from .pyramidal import L2Pyr, L5Pyr
